+++ hironaka/util/search.py
@@ -1,9 +1,7 @@
-from treelib import Node, Tree
 from typing import Tuple, List
 from collections import deque
 from dataclasses import dataclass
-from .geom import getNewtonPolytope, shift
-from .agent import AgentMorin
+from hironaka.util.geom import getNewtonPolytope, shift
 
 
 def searchDepth(points, host, debug=False):
@@ -28,7 +26,7 @@ def searchDepth(points, host, debug=False):
     return maxDepth
 
 
-def searchTree(points, tree, curr_node, host, MAX_SIZE=100):
+def searchTree(points, tree, curr_node, host, MAX_SIZE):
     """
         Perform a full tree search and store the full result in a Tree object.
     """
@@ -41,34 +39,6 @@ def searchTree(points, tree, curr_node, host, MAX_SIZE=100):
         data: List[Tuple[int]]
 
     if len(points) == 1 or tree.size() > MAX_SIZE:
-        return
-
-    shifts = []
-    coords = host.selectCoord(points)
-    for i in coords:
-        shifts.append(
-            getNewtonPolytope(
-                shift(points, coords, i)
-            )
-        )
-        node_id = tree.size()
-        tree.create_node(node_id, node_id, parent=curr_node, data=Points(shifts[-1]))
-        searchTree(shifts[-1], tree, node_id, host)
-    return tree
-
-def searchTreeMorin(points, tree, curr_node, host, MAX_SIZE=100):
-    """
-        Perform a full tree search and store the full result in a Tree object.
-    """
-
-    @dataclass
-    class Points:
-        """
-            a wrapper of a set of points.
-        """
-        data: List[Tuple[int]]
-
-    if len(points) == 1 or tree.size() > MAX_SIZE or AgentMorin.move() == False:
         return
 
     shifts = []
